Rewrite
The proposal is to let the text of non-technical introduction as long as to line 32 (with the first graphic. Then open a second site for a more technical explanation with the contents promised here. Links to the site "technical explanation" would be added in § 30 and 32. Further, the text under "Is it so bad?" may be enriched with some links to the videos posted at the "community cube" site.